+++ feature/calls/src/main/java/com/zhigaras/calls/domain/CallsCloudService.kt
@@ -1,0 +1,11 @@
+package com.zhigaras.calls.domain
+
+import com.zhigaras.calls.domain.model.ConnectionData
+import com.zhigaras.cloudeservice.CloudService
+
+interface CallsCloudService {
+    
+    fun sendToCloud(data: ConnectionData)
+    
+    fun observeUpdates(userId: String, callback: CloudService.Callback<ConnectionData>)
+}
