+++ feature/calls/src/main/java/com/zhigaras/calls/ui/CallViewModel.kt
@@ -1,7 +1,7 @@
 package com.zhigaras.calls.ui
 
 import androidx.lifecycle.viewModelScope
-import com.google.firebase.auth.FirebaseAuth
+import com.zhigaras.auth.ProvideUserId
 import com.zhigaras.calls.domain.CallCommunication
 import com.zhigaras.calls.domain.CallsController
 import com.zhigaras.calls.domain.MatchingInteractor
@@ -10,27 +10,26 @@ import com.zhigaras.core.BaseViewModel
 import com.zhigaras.core.Dispatchers
 import com.zhigaras.webrtc.databinding.FragmentCallBinding
 import kotlinx.coroutines.launch
+import org.webrtc.SurfaceViewRenderer
 
 class CallViewModel(
     private val callsController: CallsController,
     private val matchingInteractor: MatchingInteractor,
+    private val provideUserId: ProvideUserId,
     communication: CallCommunication.Mutable,
     dispatchers: Dispatchers
 ) : BaseViewModel<FragmentCallBinding, CallUiState>(communication, dispatchers) {
     
-    fun lookForOpponent(userId: String, subjectIs: String, opinion: DisputePosition) {
+    fun lookForOpponent(subjectId: String, opinion: DisputePosition) {
         viewModelScope.launch {
-//            val userId = FirebaseAuth.getInstance().uid!!
-//            val opinion = if (userId == "stmSRe5bcxNEmCTMm3bCth15vyr2") DisputePosition.SUPPORT else DisputePosition.AGAINST
-            matchingInteractor.checkMatching("1", userId, opinion).let {
+            matchingInteractor.checkMatching(subjectId, provideUserId.provide(), opinion).let {
                 it.handle(callsController, matchingInteractor)
             }
         }
     }
     
-    fun init(binding: FragmentCallBinding) {
-        callsController.initLocalView(binding.localView)
-        callsController.initRemoteView(binding.remoteView)
-        callsController.subscribeToConnectionEvents(FirebaseAuth.getInstance().uid.toString())
+    fun init(localView: SurfaceViewRenderer, remoteView: SurfaceViewRenderer) {
+        callsController.initLocalView(localView)
+        callsController.initRemoteView(remoteView)
     }
 }
