+++ feature/login/src/main/java/com/zhigaras/login/di/Module.kt
@@ -2,15 +2,16 @@ package com.zhigaras.login.di
 
 import com.zhigaras.auth.Auth
 import com.zhigaras.auth.AuthRepository
+import com.zhigaras.auth.ProvideUserId
 import com.zhigaras.login.domain.IsUserAuthorized
 import com.zhigaras.login.domain.ResetPasswordCommunication
 import com.zhigaras.login.domain.ShowId
 import com.zhigaras.login.domain.SignInCommunication
 import com.zhigaras.login.domain.SignUpCommunication
 import com.zhigaras.login.domain.UserMapper
+import com.zhigaras.login.presentation.resetpassword.ResetPasswordViewModel
 import com.zhigaras.login.presentation.signin.SignInViewModel
 import com.zhigaras.login.presentation.signup.SignUpViewModel
-import com.zhigaras.login.presentation.resetpassword.ResetPasswordViewModel
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.bind
 import org.koin.dsl.binds
@@ -43,6 +44,8 @@ fun loginModule() = module {
     )
     
     single { IsUserAuthorized.Base(get()) } bind IsUserAuthorized::class
+    
+    single { ProvideUserId.Base() } bind ProvideUserId::class
     
     single { UserMapper() }
     single { ShowId() }
